made the login modal functional
