+++ makingFitnessBetterTypescript/mfbreact/src/components/authenticationComponents/Login.tsx
@@ -6,45 +6,46 @@ import {AuthContext} from "../../dataStore";
 import { Form, Button, FormGroup, FormLabel, FormControl, FormCheck } from 'react-bootstrap';
 // import {Button, FormCheck, FormControl, FormGroup, FormLabel} from "react-bootstrap";
 
-const Login = () => {
+// @ts-ignore
+const Login = ({onSubmit ,username, setUsername, password, setPassword}) => {
 
     let auth : Auth = new Auth();
     const authContext = useContext(AuthContext);
 
-    const [username, setUsername] = useState("");
-    const [password, setPassword] = useState("");
+    // const [username, setUsername] = useState("");
+    // const [password, setPassword] = useState("");
     // const [remember, setRemember] = useState(false);
     const [error,setError]=useState(false);
     const [message,setMessage]=useState("")
     const navigate=useNavigate();
 
-    function loginUser(e:any){
-        e.preventDefault();
-        auth.login(username,password).then((response)=>{
-            console.log(response.data);
-
-            if(response.data.token){
-                localStorage.setItem("token",response.data.token);
-                localStorage.setItem("email", response.data.email);
-                localStorage.setItem("role",response.data.role);
-                authContext.setToken(response.data.token);
-            }
-
-            const currentUser = auth.getUser(username).then((response) => {
-                if(response.data){
-                    localStorage.setItem("id",response.data.memberId);
-                    localStorage.setItem("username",response.data.username);
-                }
-            });
-
-            navigate("/");
-
-        }).catch(err=>{
-            setMessage(err.response.data.exception);
-            setError(true)
-        });
-        // localStorage.setItem("remember-me",String(remember));
-    }
+    // function loginUser(e:any){
+    //     e.preventDefault();
+    //     auth.login(username,password).then((response)=>{
+    //         console.log(response.data);
+    //
+    //         if(response.data.token){
+    //             localStorage.setItem("token",response.data.token);
+    //             localStorage.setItem("email", response.data.email);
+    //             localStorage.setItem("role",response.data.role);
+    //             authContext.setToken(response.data.token);
+    //         }
+    //
+    //         const currentUser = auth.getUser(username).then((response) => {
+    //             if(response.data){
+    //                 localStorage.setItem("id",response.data.memberId);
+    //                 localStorage.setItem("username",response.data.username);
+    //             }
+    //         });
+    //
+    //         navigate("/");
+    //
+    //     }).catch(err=>{
+    //         setMessage(err.response.data.exception);
+    //         setError(true)
+    //     });
+    //     // localStorage.setItem("remember-me",String(remember));
+    // }
 
 
     // @ts-ignore
@@ -67,7 +68,7 @@ const Login = () => {
         {/*</Form>*/}
         {/*</Navbar>*/}
 
-            <Form className="login-form-container" onSubmit={loginUser}>
+            <Form className="login-form-container" onSubmit={onSubmit}>
                 <div className="login-form-background">
                     <FormGroup>
                         <FormLabel className="login-form-label"> Username</FormLabel>
